+++ tests/testthat/test-hooks-regex-exclude.R
@@ -1,13 +1,18 @@
 test_that("exclude regex for spell check hook matches expected files", {
   skip_if(not_conda())
   skip_if(on_cran())
+
+  # declare
   re <- reticulate::import("re")
-  pattern <- readLines(system.file("pre-commit-hooks.yaml", package = "precommit")) %>%
-    gsub("^ *exclude *: *>", "    exclude: |", .) %>%
-    yaml::yaml.load() %>%
-    purrr::keep(~ .x$id == "spell-check") %>%
-    magrittr::extract2(1) %>%
-    magrittr::extract2("exclude")
+
+  pattern_read <- function(file) {
+    readLines(file) %>%
+      gsub("^ *exclude *: *>", "    exclude: |", .) %>%
+      yaml::yaml.load() %>%
+      purrr::keep(~ .x$id == "spell-check") %>%
+      magrittr::extract2(1) %>%
+      magrittr::extract2("exclude")
+  }
 
   is_match <- function(pattern, x) {
     matches <- purrr::map_lgl(
@@ -42,8 +47,16 @@ test_that("exclude regex for spell check hook matches expected files", {
     "more/data.rds",
     "things/xx.Rds",
     ".pre-commit-",
-    ".Rproj"
+    ".Rproj",
+    "some/X.pdf",
+    "figure.png",
+    "pciture.jpeg",
+    ".github/workflows/r-cmd-check.yaml"
   )
+
+  # run
+  pattern <- pattern_read(system.file("pre-commit-hooks.yaml", package = "precommit"))
+
   individual_patterns <- setdiff(
     unlist(strsplit(pattern, "\n", fixed = TRUE)),
     c("(?x)^(", ")$")